# dev/openapi.yaml
openapi: 3.1.0
info:
  title: Chronix REST API
  version: 0.1.0
  description: |
    Chronix is a zero-dependency job scheduler that runs SQL tasks on time.

    This OpenAPI document reflects the current server endpoints and data types as implemented in the repo.

    Conventions:
    - All JSON properties use camelCase.
    - Responses use `restresponse` helpers: success returns raw objects; errors return a standard error envelope.
  contact:
    name: Chronix
  license:
    name: ISC
servers:
  - url: http://localhost:5170
    description: Local development
  - url: /
    description: Relative to the current origin (production)
security:
  - cookieAuth: []
...
paths:
...
  /connections/{id}/test:
    post:
      tags: [DB Connections]
      summary: Test connection by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DbTestResult'
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
  /connections/test:
    post:
      tags: [DB Connections]
      summary: Test draft connection payload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DbConnectionPayload'
      responses:
        '200':
          description: Test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DbTestResult'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cxtoken
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthenticated:
      description: Unauthenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Stable application error code (restresponse Code enum)
          example: BadRequest
        message:
          type: string
        details:
          description: Optional error details
          oneOf:
            - type: array
              items: { type: string }
            - type: object
      required: [code, message]
...
    DbConnectionPayload:
      type: object
      properties:
        name: { type: string }
        driver: { type: string, enum: [postgres, sqlite] }
        dsn: { type: string }
        description: { type: string, nullable: true }
        enabled: { type: boolean, nullable: true }
        autoCheckEnabled: { type: boolean, nullable: true }
        autoCheckSeconds: { type: integer, format: int64, nullable: true }
        agentUuid: { type: string, format: uuid, nullable: true }
      required: [name, driver, dsn]
...
    DbTestResult:
      type: object
      properties:
        ok: { type: boolean }
        error: { type: string, nullable: true }
        latencyMs: { type: integer, format: int64, nullable: true }
...